[ao] Removing memoryless observer args for MovingAverage
The original implementation of memoryless observers used MinMaxObservers and
a memoryless argument to manipulate the behavior of the observer such that it wouldn't
keep track of previously observed min and max's. It was later pointed
out that this was equivalent to a movingaverageobserver with averaging_constant=1
which is requires less overhead and no 1 off args (memoryless) so this PR refactors
the memoryless arg and uses MovingAverage observers instead, although the memoryless
adjective is still used, a complete definintion was also added to clarify error
messages given these changes.

ghstack-source-id: 49ff050e45747c92fc4bc1bc2a6eceb165b46864
Pull Request resolved: https://github.com/pytorch/pytorch/pull/73947